+++ src/shapes/circle.js
@@ -1,4 +1,4 @@
-import _ from "underscore"
+import assign from "lodash/object/assign"
 import { Moveable, Styleable } from "../mixins"
 import Ellipse from "./ellipse"
 import Utils from '../utils'
@@ -33,6 +33,6 @@ class Circle {
 }
 
 // Should we figure out a better way to do mixins for ES6?
-_.extend(Circle.prototype, Moveable, Styleable, { type: "circle" });
+assign(Circle.prototype, Moveable, Styleable, { type: "circle" });
 
 export default Circle;