+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 0.2.9
 
 - `Rune.Polygon` now has a `contains(x,y)` function to know whether a point is inside or outside a polygon
-- Group and shapes now has a `stagepos()` function that returns their absolute position according to the stage
+- Group and shapes now have a `stagepos()` function that returns their absolute position according to the stage
 - Removing `Groupable` mixin in favor of `Moveable`
+- Replacing underscore with lodash to shrink size of library
 
 ## 0.2.8